# Inlining everything