--- a/src/Parameters.jl
+++ b/src/Parameters.jl
@@ -56,20 +56,20 @@ Base.eltype(::ThermodynamicsParameters{FT}) where {FT} = FT
 
 # wrappers
 for fn in fieldnames(ATP)
-    @eval $(fn)(ps::ATP) = ps.$(fn)
+    @eval @inline $(fn)(ps::ATP) = ps.$(fn)
 end
 
 # Derived parameters
-R_d(ps::ATP) = ps.gas_constant / ps.molmass_dryair
-R_v(ps::ATP) = ps.gas_constant / ps.molmass_water
-molmass_ratio(ps::ATP) = ps.molmass_dryair / ps.molmass_water
-LH_f0(ps::ATP) = ps.LH_s0 - ps.LH_v0
-e_int_v0(ps::ATP) = ps.LH_v0 - R_v(ps) * ps.T_0
-e_int_i0(ps::ATP) = LH_f0(ps)
-cp_d(ps::ATP) = R_d(ps) / ps.kappa_d
-cv_d(ps::ATP) = cp_d(ps) - R_d(ps)
-cv_v(ps::ATP) = ps.cp_v - R_v(ps)
-cv_l(ps::ATP) = ps.cp_l
-cv_i(ps::ATP) = ps.cp_i
+@inline R_d(ps::ATP) = ps.gas_constant / ps.molmass_dryair
+@inline R_v(ps::ATP) = ps.gas_constant / ps.molmass_water
+@inline molmass_ratio(ps::ATP) = ps.molmass_dryair / ps.molmass_water
+@inline LH_f0(ps::ATP) = ps.LH_s0 - ps.LH_v0
+@inline e_int_v0(ps::ATP) = ps.LH_v0 - R_v(ps) * ps.T_0
+@inline e_int_i0(ps::ATP) = LH_f0(ps)
+@inline cp_d(ps::ATP) = R_d(ps) / ps.kappa_d
+@inline cv_d(ps::ATP) = cp_d(ps) - R_d(ps)
+@inline cv_v(ps::ATP) = ps.cp_v - R_v(ps)
+@inline cv_l(ps::ATP) = ps.cp_l
+@inline cv_i(ps::ATP) = ps.cp_i
 
 end