This PR attempts to systematically inline all thermo functions, to see the impact of performance.