@charleskawczynski 2024-01-22
I took a look at the buoyancy kernel, and I'm seeing a speedup on this branch (on the CPU):

Main branch
```julia
******************* thermo_kernel
BenchmarkTools.Trial: 236 samples with 1 evaluation.
 Range (min … max):  20.429 ms …  28.884 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     20.910 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   21.210 ms ± 876.287 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▄▃█▆▁                                                      
  ▃▄███████▅▄▃▄▅▂▃▄▃▁▂▂▁▂▃▃▄▃▄▂▂▃▂▃▃▃▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▃
  20.4 ms         Histogram: frequency by time         24.8 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
******************* buoyancy_gradients_kernel
BenchmarkTools.Trial: 27 samples with 1 evaluation.
 Range (min … max):  180.741 ms … 196.250 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     183.876 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   185.915 ms ±   4.569 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▃    ▃   ██▃                                                  
  ▇█▁▁▇▁█▁▁▁███▇▇▇▇▁▁▇▁▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▁▇▁▁▇▁▇▁▁▁▁▁▁▁▇▁▁▁▁▁▇▁▇ ▁
  181 ms           Histogram: frequency by time          196 ms <

 Memory estimate: 16 bytes, allocs estimate: 1.
```

This branch:
```julia
******************* thermo_kernel
BenchmarkTools.Trial: 368 samples with 1 evaluation.
 Range (min … max):  13.501 ms …  14.498 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     13.544 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   13.592 ms ± 143.209 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁██▇▆▅▆▄▃▃▂▁                                                  
  ████████████▇▆▁▁▄▄▁▁▁▆▁▇▄▁▆▁▁▆▄▁▁▁▁▄▄▁▄▁▄▄▁▁▁▄▁▁▁▁▁▁▄▆▆▇▆▄▄▆ ▇
  13.5 ms       Histogram: log(frequency) by time      14.1 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
******************* buoyancy_gradients_kernel
BenchmarkTools.Trial: 36 samples with 1 evaluation.
 Range (min … max):  138.371 ms … 142.465 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     139.539 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   139.659 ms ± 890.868 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

               ▂   █▂  ▂                                         
  ▅▁▅▁▅▅█▅▅▁▅▅▁█▅▅▅██▁▁█▁▅▅▁▅▅▁▁▁▁▁▁▁▅▁▅▁▁▁▁▁▁▁▁▅▁▁▁▅▁▁▁▁▁▁▁▁▁▅ ▁
  138 ms           Histogram: frequency by time          142 ms <

 Memory estimate: 16 bytes, allocs estimate: 1.
```

I'm not sure if the input parameters are performing saturation adjustment, but it doesn't matter in the case of the buoyancy gradients kernel, since SA is never called regardless of what the inputs are.

cc @glwagner, @szy21. I'm comfortable moving forward with this.

Interestingly, I didn't realize this until now, it looks like the buoyancy