@trontrytel 2024-01-22
This looks great! So maybe it will speed up the quadratures also...